🎨 Added custom hook to get horses
