--- client/src/components/loader/Loader.tsx
@@ -1,9 +1,0 @@
-import { PulseLoader } from "react-spinners";
-
-export function Loader() {
-    return(
-        <>
-            <PulseLoader color="#0A3B2C" speedMultiplier={0.7}></PulseLoader>
-        </>
-    )
-}
