+++ client/src/components/loader/ShowLoader.tsx
@@ -1,0 +1,9 @@
+import { PulseLoader } from "react-spinners";
+
+export function ShowLoader() {
+    return(
+        <>
+            <PulseLoader color="#0A3B2C"></PulseLoader>
+        </>
+    )
+}
